refactoring: removed old code for profile button in translator and translator button in profile. Merged localstorage files into one local storage file.
